feat(boot-options): changes the options interface of the Microservice.
The microservice constructor now takes an options object. While this object previously contained the boot options for Loopback, these configuration is now moved to the boot property of the options object.

BREAKING CHANGE: Directly passing the boot options to the constructor and the helper methods to boot and start the service will no longer work.